in the process of development
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@53 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/payment.h
@@ -337,7 +337,7 @@
 
 typedef struct {
 
-                Uint16          gold;
+                Uint32          gold;
                 Uint16          wood;
                 Uint16          ore;
                 Uint16          mercury;
